# Optimize `zip_findlast_entry` and add benchmarks