--- a/benchmark/README.md
+++ b/benchmark/README.md
@@ -1,0 +1,11 @@
+# ZipArchives.jl benchmarks
+
+This directory contains benchmarks for ZipArchives. To run all the
+benchmarks, launch `julia --project=benchmark` and enter:
+
+``` julia
+using PkgBenchmark
+import ZipArchives
+
+benchmarkpkg(ZipArchives)
+```