|                                    | main               | dirty              | main/dirty |
|:-----------------------------------|:------------------:|:------------------:|:----------:|
| reading/ZipReader                  | 0.0427 ± 0.0026 ms | 0.0429 ± 0.0027 ms | 0.995      |
| reading/zip_findlast_entry first   | 7.02 ± 0.1 μs      | 7.45 ± 0.08 μs     | 0.942      |
| reading/zip_findlast_entry last    | 0.08 ± 0.01 μs     | 0.07 ± 0.01 μs     | 1.14       |
| reading/zip_findlast_entry nothing | 5.47 ± 0.27 μs     | 1.15 ± 0.04 μs     | 4.75       |
| reading/zip_readentry              | 5.62 ± 1.2 μs      | 5.59 ± 1.2 μs      | 1.01       |
| time_to_load                       | 0.0559 ± 0.00035 s | 0.0574 ± 0.00036 s | 0.974      |